refactor(cl): convert RECAPSearchTestCase to mixin
