+++ cl/lib/test_helpers.py
@@ -15,10 +15,6 @@ from cl.audio.factories import AudioFactory
 from cl.audio.models import Audio
 from cl.lib.utils import deepgetattr
 from cl.people_db.factories import (
-    AttorneyFactory,
-    AttorneyOrganizationFactory,
-    PartyFactory,
-    PartyTypeFactory,
     PersonFactory,
 )
 from cl.search.constants import o_type_index_map
@@ -26,17 +22,11 @@ from cl.search.docket_sources import DocketSources
 from cl.search.documents import DocketDocument
 from cl.search.factories import (
     CourtFactory,
-    DocketEntryWithParentsFactory,
     DocketFactory,
-    OpinionClusterFactory,
-    OpinionFactory,
-    RECAPDocumentFactory,
 )
 from cl.search.models import (
     Citation,
     Docket,
-    OpinionsCitedByRECAPDocument,
-    RECAPDocument,
 )
 from cl.tests.cases import TestCase
 from cl.users.factories import UserProfileWithParentsFactory
@@ -781,119 +771,6 @@ audio_v4_fields.update(
         "meta": [],  # type: ignore
     }
 )
-
-
-class RECAPSearchTestCase(TestCase):
-    """RECAP Search test case factories"""
-
-    @classmethod
-    def setUpTestData(cls):
-        cls.court = CourtFactory(id="canb", jurisdiction="FB")
-        cls.court_2 = CourtFactory(id="ca1", jurisdiction="F")
-        cls.judge = PersonFactory.create(
-            name_first="Thalassa", name_last="Miller"
-        )
-        cls.judge_2 = PersonFactory.create(
-            name_first="Persephone", name_last="Sinclair"
-        )
-        cls.de = DocketEntryWithParentsFactory(
-            docket=DocketFactory(
-                court=cls.court,
-                case_name="SUBPOENAS SERVED ON",
-                case_name_full="Jackson & Sons Holdings vs. Bank",
-                date_filed=datetime.date(2015, 8, 16),
-                date_argued=datetime.date(2013, 5, 20),
-                docket_number="1:21-bk-1234",
-                assigned_to=cls.judge,
-                referred_to=cls.judge_2,
-                nature_of_suit="440",
-                source=Docket.RECAP,
-                cause="401 Civil",
-                jurisdiction_type="'U.S. Government Defendant",
-                jury_demand="1,000,000",
-            ),
-            entry_number=1,
-            date_filed=datetime.date(2015, 8, 19),
-            description="MOTION for Leave to File Amicus Curiae Lorem Served",
-        )
-        cls.firm = AttorneyOrganizationFactory(name="Associates LLP")
-        cls.attorney = AttorneyFactory(
-            name="Debbie Russell",
-            organizations=[cls.firm],
-            docket=cls.de.docket,
-        )
-        cls.party_type = PartyTypeFactory.create(
-            party=PartyFactory(
-                name="Defendant Jane Roe",
-                docket=cls.de.docket,
-                attorneys=[cls.attorney],
-            ),
-            docket=cls.de.docket,
-        )
-
-        cls.rd = RECAPDocumentFactory(
-            docket_entry=cls.de,
-            description="Leave to File",
-            document_number="1",
-            is_available=True,
-            page_count=5,
-            pacer_doc_id="018036652435",
-        )
-
-        cls.opinion = OpinionFactory(
-            cluster=OpinionClusterFactory(docket=cls.de.docket)
-        )
-        OpinionsCitedByRECAPDocument.objects.bulk_create(
-            [
-                OpinionsCitedByRECAPDocument(
-                    citing_document=cls.rd,
-                    cited_opinion=cls.opinion,
-                    depth=1,
-                )
-            ]
-        )
-        cls.rd_att = RECAPDocumentFactory(
-            docket_entry=cls.de,
-            description="Document attachment",
-            document_type=RECAPDocument.ATTACHMENT,
-            document_number="1",
-            attachment_number=2,
-            is_available=False,
-            page_count=7,
-            pacer_doc_id="018036652436",
-        )
-
-        cls.judge_3 = PersonFactory.create(
-            name_first="Seraphina", name_last="Hawthorne"
-        )
-        cls.judge_4 = PersonFactory.create(
-            name_first="Leopold", name_last="Featherstone"
-        )
-        cls.de_1 = DocketEntryWithParentsFactory(
-            docket=DocketFactory(
-                docket_number="12-1235",
-                court=cls.court_2,
-                case_name="SUBPOENAS SERVED OFF",
-                case_name_full="The State of Franklin v. Solutions LLC",
-                date_filed=datetime.date(2016, 8, 16),
-                date_argued=datetime.date(2012, 6, 23),
-                assigned_to=cls.judge_3,
-                referred_to=cls.judge_4,
-                source=Docket.COLUMBIA_AND_RECAP,
-            ),
-            entry_number=3,
-            date_filed=datetime.date(2014, 7, 5),
-            description="MOTION for Leave to File Amicus Discharging Debtor",
-        )
-        cls.rd_2 = RECAPDocumentFactory(
-            docket_entry=cls.de_1,
-            description="Leave to File",
-            document_number="3",
-            page_count=10,
-            plain_text="Mauris iaculis, leo sit amet hendrerit vehicula, Maecenas nunc justo. Integer varius sapien arcu, quis laoreet lacus consequat vel.",
-            pacer_doc_id="016156723121",
-        )
-        super().setUpTestData()
 
 
 class SimpleUserDataMixin:
